# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Task Manager API",
    "description" : "REST foundation for the Task Manager project",
    "version" : "v1"
  },
  "tags" : [ {
    "name" : "Actuator",
    "description" : "Monitor and interact",
    "externalDocs" : {
      "description" : "Spring Boot Actuator Web API Documentation",
      "url" : "https://docs.spring.io/spring-boot/docs/current/actuator-api/html/"
    }
  }, {
    "name" : "Application",
    "description" : "Versioned application foundation endpoints"
  }, {
    "name" : "Tasks",
    "description" : "Task management endpoints"
  } ],
  "paths" : {
    "/api/v1/tasks/{taskId}" : {
      "get" : {
        "tags" : [ "Tasks" ],
        "summary" : "Get a task by ID",
        "description" : "Returns a persisted task by its unique identifier.",
        "operationId" : "getTaskById",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Task found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid task identifier",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404" : {
            "description" : "Task not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Tasks" ],
        "summary" : "Replace a task",
        "description" : "Replaces a task's title, description, and status.",
        "operationId" : "updateTask",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Task updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404" : {
            "description" : "Task not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Tasks" ],
        "summary" : "Delete a task",
        "description" : "Permanently deletes a task by its unique identifier.",
        "operationId" : "deleteTask",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Task deleted"
          },
          "400" : {
            "description" : "Invalid task identifier",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404" : {
            "description" : "Task not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks" : {
      "get" : {
        "tags" : [ "Tasks" ],
        "summary" : "List tasks",
        "description" : "Returns a paginated list of tasks with optional filtering and sorting.",
        "operationId" : "listTasks",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Tasks found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskListResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request parameters",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Tasks" ],
        "summary" : "Create a task",
        "description" : "Creates a task with server-managed status and timestamps.",
        "operationId" : "createTask",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Task created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/status" : {
      "patch" : {
        "tags" : [ "Tasks" ],
        "summary" : "Update a task's status",
        "description" : "Changes only a task's lifecycle status.",
        "operationId" : "updateTaskStatus",
        "parameters" : [ {
          "name" : "taskId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTaskStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Task status updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404" : {
            "description" : "Task not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1" : {
      "get" : {
        "tags" : [ "Application" ],
        "summary" : "Get application identity",
        "description" : "Returns the service name and current API version.",
        "operationId" : "getApplicationInfo",
        "responses" : {
          "200" : {
            "description" : "Application identity",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplicationResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Unsupported HTTP method",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/actuator" : {
      "get" : {
        "tags" : [ "Actuator" ],
        "summary" : "Actuator root web endpoint",
        "operationId" : "links",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/vnd.spring-boot.actuator.v3+json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object",
                    "additionalProperties" : {
                      "$ref" : "#/components/schemas/Link"
                    }
                  }
                }
              },
              "application/vnd.spring-boot.actuator.v2+json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object",
                    "additionalProperties" : {
                      "$ref" : "#/components/schemas/Link"
                    }
                  }
                }
              },
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "object",
                    "additionalProperties" : {
                      "$ref" : "#/components/schemas/Link"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/actuator/health" : {
      "get" : {
        "tags" : [ "Actuator" ],
        "summary" : "Actuator web endpoint 'health'",
        "operationId" : "health",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/vnd.spring-boot.actuator.v3+json" : {
                "schema" : {
                  "type" : "object"
                }
              },
              "application/vnd.spring-boot.actuator.v2+json" : {
                "schema" : {
                  "type" : "object"
                }
              },
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "UpdateTaskRequest" : {
        "type" : "object",
        "description" : "Task replacement request",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "Task title",
            "example" : "Prepare release notes",
            "minLength" : 1
          },
          "description" : {
            "type" : "string",
            "description" : "Task description",
            "example" : "Summarize the changes for the next release"
          },
          "status" : {
            "type" : "string",
            "description" : "Task lifecycle status",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ],
            "example" : "IN_PROGRESS"
          }
        },
        "required" : [ "description", "status", "title" ]
      },
      "TaskResponse" : {
        "type" : "object",
        "description" : "Persisted task",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "3f6f0f2e-3f2c-4b61-8c1b-0b3d4c5d6e7f"
          },
          "title" : {
            "type" : "string",
            "example" : "Prepare release notes"
          },
          "description" : {
            "type" : "string",
            "example" : "Summarize the changes for the next release"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ],
            "example" : "TODO"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-01-01T12:00:00Z"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-01-01T12:00:00Z"
          }
        }
      },
      "ApiError" : {
        "type" : "object",
        "description" : "Standard error returned by the REST API",
        "properties" : {
          "errorId" : {
            "type" : "string",
            "example" : "METHOD_NOT_ALLOWED"
          },
          "message" : {
            "type" : "string",
            "example" : "HTTP method is not supported for this resource"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-01-01T12:00:00Z"
          },
          "correlationId" : {
            "type" : "string",
            "example" : "3f6f0f2e-3f2c-4b61-8c1b-0b3d4c5d6e7f"
          }
        }
      },
      "CreateTaskRequest" : {
        "type" : "object",
        "description" : "Task creation request",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "Task title",
            "example" : "Prepare release notes",
            "minLength" : 1
          },
          "description" : {
            "type" : "string",
            "description" : "Task description",
            "example" : "Summarize the changes for the next release"
          }
        },
        "required" : [ "description", "title" ]
      },
      "UpdateTaskStatusRequest" : {
        "type" : "object",
        "description" : "Task status update request",
        "properties" : {
          "status" : {
            "type" : "string",
            "description" : "Task lifecycle status",
            "enum" : [ "TODO", "IN_PROGRESS", "DONE" ],
            "example" : "IN_PROGRESS"
          }
        },
        "required" : [ "status" ]
      },
      "ApplicationResponse" : {
        "type" : "object",
        "description" : "Identity and API version of the running service",
        "properties" : {
          "service" : {
            "type" : "string",
            "example" : "task-manager"
          },
          "version" : {
            "type" : "string",
            "example" : "v1"
          }
        }
      },
      "TaskListResponse" : {
        "type" : "object",
        "description" : "Paged task listing response",
        "properties" : {
          "items" : {
            "type" : "array",
            "description" : "Task items on this page",
            "items" : {
              "$ref" : "#/components/schemas/TaskResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 20
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64",
            "example" : 150
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 8
          }
        }
      },
      "Link" : {
        "type" : "object",
        "properties" : {
          "href" : {
            "type" : "string"
          },
          "templated" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}